Enable GPU Memory profiler tests (#18701)
* Enable GPU Memory profiler tests

Previously tests are not run as test_profiler.py was not taken into account on
GPU CI runs and some tests were marked for being skipped if run on a CPU-only
machine.

* Disable broken tests
